Display Price with Tax directly
When we write price without taxe, the other field ( Need to be activated by changing $inputalsopricewithtax = 0 to 1 ) "Price with Tax" display the price with tax and vice-versa when we write in the field "Price with Tax" the field "Price without Tax" display the price without price.

Before this change, when we wrote something in one of those field, the other was erased. 

PS: I'm french so the description can contains some mistakes.